backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart==0.0.6
firebase-admin==6.2.0
pdfplumber
requests
beautifulsoup4
lxml
langchain-text-splitters
google-genai
python-dotenv
apscheduler
huggingface_hub
